refactor: extract textarea completions component
